requirements.txt
Flask==2.0.2
Flask-MySQLdb==0.2.0
mysqlclient==2.1.0